% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ro_desc.R
\name{add_ro_desc}
\alias{add_ro_desc}
\title{Add rOpenSci review mention to DESCRIPTION}
\usage{
add_ro_desc(path = usethis::proj_get())
}
\arguments{
\item{path}{path to package}
}
\description{
Add rOpenSci review mention to DESCRIPTION. Only run this
after your package has been approved.
}
