% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ro_fnd.R
\name{add_ro_fnd}
\alias{add_ro_fnd}
\title{Add rOpenSci as funder to DESCRIPTION}
\usage{
add_ro_fnd(path = usethis::proj_get())
}
\arguments{
\item{path}{path to package}
}
\description{
Add rOpenSci as funder to DESCRIPTION
}
