% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blog_posts.R
\name{create_post_project}
\alias{create_post_project}
\title{Create a RStudio project with a blog post template}
\usage{
create_post_project(slug, type = "Rmd")
}
\arguments{
\item{slug}{Slug describing the blog post, e.g
"thanking-reviewers-in-metadata". Use hyphens, not space.}

\item{type}{Either 'Rmd' or 'md'}
}
\value{
It opens a new project.
}
\description{
Create a RStudio project with a blog post template
}
