% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optout_rodocs.R
\name{optout_rodocs}
\alias{optout_rodocs}
\title{Opt out of rOpenSci docs website building}
\usage{
optout_rodocs(path = usethis::proj_get())
}
\arguments{
\item{path}{path to package}
}
\value{
It creates an empty .norodocs file,
nothing else has to be done.
}
\description{
Opt out of rOpenSci docs website building
}
\details{
Run this function if you do not want your
package to get a website built and deployed at \code{docs.ropensci.org/pkg_name}.
}
