% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blog_posts.R
\name{prepare_fork}
\alias{prepare_fork}
\title{Prepare fork}
\usage{
prepare_fork(post_project_path, gh_username = "maelle")
}
\arguments{
\item{post_project_path}{Path to the project where the post is.}

\item{gh_username}{GitHub username.}
}
\description{
Based on project with post.md and images in images/,
forks the website repo, put the post content in the right place,
the images in the right place and correct the path to them in the
post.
}
\details{
This relies on your having a GitHub PAT.
}
