% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{repostatus_badges}
\alias{repostatus_badges}
\title{Names and Markdown code of repostatus.org badges}
\format{
A data.frame with columns status (lowercase) and md_code.
}
\source{
\url{http://www.repostatus.org/}
}
\description{
Names, Markdown code, href and src of repostatus.org badges
as retrieved from \url{https://github.com/jantman/repostatus.org/tree/master/badges/latest}
before this package release. See also \url{http://www.repostatus.org/}
The data is distributed with a CC-BY SA license. Refer to \url{http://www.repostatus.org/}
}
