% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.R
\name{prepare_submission}
\alias{prepare_submission}
\alias{post_submission}
\title{Prepare an rOpenSci submission}
\usage{
prepare_submission(pkg = usethis::proj_get())

post_submission(pkg = usethis::proj_get())
}
\arguments{
\item{pkg}{Path to the package source}
}
\description{
Prepare an rOpenSci submission
}
\details{
Workflow
\itemize{
\item \code{prepare_submission()}: Create a Markdown file for rOpenSci submission, and pre-fills it.
\item Edit the ropensci-submission.md file, put it under version control if desired.
\item \code{post_submission()}: Puts the submission info on the clipboard and open a GitHub page.
}
}
