% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_cchecks_badge.R
\name{use_cchecks_badge}
\alias{use_cchecks_badge}
\title{Use CRAN checks badge}
\usage{
use_cchecks_badge()
}
\value{
It will help you add the CRAN checks badge to your README.
}
\description{
Use CRAN checks badge
}
\examples{
\dontrun{
use_cchecks_badge()
}
}
