% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_review_badge.R
\name{use_review_badge}
\alias{use_review_badge}
\title{Use rOpenSci review badge}
\usage{
use_review_badge(onboarding_issue_id)
}
\arguments{
\item{onboarding_issue_id}{Issue number of the onboarding thread of your package}
}
\value{
It will help you add the repostatus.org badge to your README.
}
\description{
Use rOpenSci review badge
}
\examples{
use_review_badge(24)
}
