% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_ro_aut_pkgdown.R
\name{use_ro_aut_pkgdown}
\alias{use_ro_aut_pkgdown}
\title{Generate YAML to add rOpenSci as author in pkgdown config}
\usage{
use_ro_aut_pkgdown()
}
\description{
Generate YAML to add rOpenSci as author in pkgdown config
}
