% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown.R
\name{use_ro_favicon}
\alias{use_ro_favicon}
\title{Use rOpenSci favicon for pkgdown}
\usage{
use_ro_favicon()
}
\description{
Use rOpenSci favicon for pkgdown
}
\details{
This will create a pkgdown folder with the favicon files.
}
