% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaboration_templates.R
\name{use_ro_github}
\alias{use_ro_github}
\title{Add templates in .github}
\usage{
use_ro_github(package_name)
}
\arguments{
\item{package_name}{Package name}
}
\description{
Add templates in .github
}
