% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_repostatus_badge.R
\name{use_repostatus_badge}
\alias{use_repostatus_badge}
\title{Use repostatus.org Badge}
\usage{
use_repostatus_badge(status)
}
\arguments{
\item{status}{current status of the project, cf \url{repostatus.org}}
}
\value{
It will help you add the repostatus.org badge to your README.
}
\description{
Use repostatus.org Badge
}
\examples{
\dontrun{
use_repostatus_badge("wip")
}
}
